GetRootNode.js First Upload
Added the first fully documented version of a new, cross-browser polyfill function, getRootNodePolyfill() in GetRootNode.js, in the "ancestry" folder. getRootNodePolyfill() calls the Node interface's getRootNode() method, if it is supported, to get a root Node, a Node from which all others in a hierarchy are children and has no parents itself, for a target Node object. Otherwise, it traverses up a DOM tree, Node parent, or HTML ancestor hierarchy until either the true root Node with no further ancestors is found or the first associated ShadowRoot to the target Node is found, depending on the value of the "options.composed" parameter passed to it. "Ancestors" are defined as Node.parentNode entries or ShadowRoot "host" property nodes. The getRootNodePolyfill() function overrides the Node's interface's getRootNode() method in its prototype if it is not supported by an Internet browser. Since the getRootNode() property is relatively new, this polyfill function is extremely useful for legacy Internet browsers, such as Internet Explorer and Microsoft Spartan (non-Chromium) Edge. ShadowRoots are ignored for browsers which do not support them.