GetOwnerDoc.js First Upload
Added the first fully-documented version of a cross-browser, polyfill function called getOwnerDocument() in the GetOwnerDoc.js file. The getOwnerDocument() function aims to return a target Node object's owner document, which is the Document object which created it. It calls the Node interface's "ownerDocument" property, if supported, to return the owner Document object, if available. If the "ownerDocument" property is not supported, the function traverses up the ancestral Node hierarchy or DOM tree structure to which the Node is attached using the "parentNode" property until a Node with no further parent, called the root Node, is found. It tests whether this root Node is a Document object, and if it is not, then it returns null. This function also returns null if a Document node is passed to it, in keeping with the behavior of the "ownerDocument" property.