ReplaceNode.js First Upload
Added the first, partially documented version of a new, cross-browser, polyfill function, called replaceNode(), in the replace/ReplaceNode.js file. replaceNode() takes in a single Node object or a list of Node objects, either already contained in an HTML document's DOM structure or generated by JavaScript DOM-building methods, and replaces them with a single Node object, Node object(s) parsed from a string with HTML markup, or a list of such entries, which also may or may not be contained in an existing HTML document's DOM structure. 

Not only does this function call newer or legacy JavaScript DOM replacement and string parsing methods when appropriate, but it also covers cases which traditional JavaScript replacement methods do not, such as replacing the entire child Node contents of Document or DocumentFragment objects. There are several options available for modifying the behavior of the function: "cloneNodes" to make deep copies of replacement objects before replacing the originals with them; "getNewList" to return a list of the nodes that either replaced or became child Nodes of the originals; and "replaceOnce" to specify whether Nodes should be replaced once in the order that they occur for both the original and replacement entries. By default, the function will replace a single, original Node value passed to the function using a given replacement object WITHOUT making deep copies; it will return the number of original Nodes which were replaced; and it will replace each original entry with a copy of each replacement Node passed to the function in a list. 

I'm still working on adding features to the replaceNode() function, and there are many different possible input cases that make it a powerful utility function. The function has general, extensive documentation covering how to use it at its top, but its member methods and code blocks are not fully documented yet. I've written and ran as many input and test cases for the function as I could think of. However, there may still be bugs or input cases which produce exceptions which I may not have have thought of or ran yet, so please let know if you encounter any.