RemoveChildren.js First Upload
Added the first, fully-documented version of a new, cross-browser, client-side polyfill and general utility function called removeChildren(), in the directory and file at remove/RemoveChildren.js. This function removes all of the child Nodes for a specified parent Node object. Though written using ES3 syntax for maximum Internet browser compatibility, the function prioritizes the latest and most optimized JavaScript APIs and faster algorithms for removing child Nodes over legacy ones, which tend to be slower. Optionally, only child Nodes which are Element objects may be removed, while the child Nodes of other types will be kept. This may be useful when a programmer may have reasons to remove either visible or hidden Element children from a parent Node, but not other child Node content, such as comment, processing instruction, or text Nodes.

Also optionally, child Nodes can be removed in either an ascending, "forward" direction or ordering, from the first to the last Node; or in a descending, "backwards" direction or ordering, from the last to the first Node. By default, if not specified, child Nodes will be removed all at once or in either direction which provides the best performance for certain JavaScript APIs, properties, or methods for child Node removal. If there are lots of visible child Nodes attached to the target parent Node, removing child Nodes in the "forward" direction can give the illusion of an downwards collapsing effect, and removing child Nodes in the "reverse" direction may appear as an upwards collapsing effect. However, the default DOM node removal process is usually so quick that it will not substitute for animations, but this function can be used quite well alongside animation functions for removing hidden Elements on a page. Additionally, depending on whether a user's Internet browser viewport is above or below the target parent Node on a page, any jumping or jarring of a user's viewport position resulting from content suddenly being removed from a page can be minimized by removing child Nodes in the opposite direction of their viewport positioning, relative to the target parent Node.