First GetAllComments.js Upload
Added the first, fully-documented version of a new general utility function, getAllComments() in GetAllComments.js, intended to be widely compatible with both legacy and modern browsers. This function gets all of the Comment Nodes inside of a containing Node's hierarchical DOM tree descendant structure. JavaScript APIs, methods, and traversal approaches which are more modern and offer better DOM tree traversal performance are prioritized over ones which accommodate legacy browsers, are more widely compatible, but which are slower in performance. The latter approaches are used as fallbacks for legacy Internet browsers.

Optionally, it can only get Comment Nodes that contain a string search phrase in their text contents. By default, the search is case-insensitive, e.g. "search" is matched the same as "seArCH", and loosely matches any sequences, including parts of words, in a Comment Node's text contents. However, it can also optionally be instructed to match a search phrase case-sensitively, e.g. "search" is not the same as "SEARCH",  and/or only match whole-word sequences, e.g. "search" will be matched in "search phrase" but not in "searchPhrase". This function is particularly useful for scripts and web pages which rely on Comment Nodes as invisible placeholders for content that will later be inserted into the web page, or which hide the content that will later be inserted into the web page inside of a Comment Node's text contents.