array/At.js First Upload
Added the first, fully documented version of a new polyfill function for the Array class's at() method, a relatively new and experimental method in ESNext JavaScript for getting the element at an index in an Array or Array-like object. At the time of uploading this file, the Array's at() method is not widely supported by the official versions of mainstream Internet browsers and is only supported in some beta or developer preview builds of them. This function can thus be used for polyfilling the at() method until is is more widely implemented by Internet browsers or by developers who want to use the at() method in older Internet browsers.